Separate KavenegarApi return models
